feat(boom): B1 -- load ZDBSP extended nodes (XNOD)
Boom map compat, step 1: the loading gate for large modern maps. aiDoom's node/subsector fields
were 16-bit; upgrade node_t.children and subsector_t.numlines/firstline to int, and move
NF_SUBSECTOR from the on-disk bit-15 (0x8000, now NF_SUBSECTOR_16) to an in-memory bit-31
(0x80000000). P_LoadNodes converts vanilla children accordingly; the BSP traversal (r_bsp/r_main/
p_sight) uses NF_SUBSECTOR symbolically so it needs no change.

Add P_LoadNodes_Extended: when the NODES lump starts with "XNOD" (SSECTORS/SEGS empty) parse the
whole extended tree from it -- appended vertices, subsectors (implicit firstseg), segs (32-bit
v1/v2 with computed angle + offset), and 32-bit-children nodes -- replacing the three vanilla
loaders. P_SetupLevel falls back to the vanilla path when it's not XNOD.

Verified: vanilla DOOM renders unchanged; an e1-arenas map rebuilt with `zdbsp -X` loads
("extended (XNOD) nodes -- 569 verts, 307 ssectors, 1021 segs, 306 nodes") and renders identically.
ZNOD (zlib-compressed) + GL variants are not handled yet.

+++ files/r_defs.h
@@ -227,8 +227,8 @@ typedef struct line_s
 typedef struct subsector_s
 {
     sector_t*	sector;
-    short	numlines;
-    short	firstline;
+    int		numlines;	// int for extended nodes (Boom/ZDBSP)
+    int		firstline;
     
 } subsector_t;
 
@@ -273,8 +273,8 @@ typedef struct
     // Bounding box for each child.
     fixed_t	bbox[2][4];
 
-    // If NF_SUBSECTOR its a subsector.
-    unsigned short children[2];
+    // If NF_SUBSECTOR its a subsector.  int (not short) for Boom/ZDBSP extended nodes.
+    int children[2];
     
 } node_t;
